Moved 'requiredjson' library in his own package
