+++ go.mod
@@ -57,7 +57,10 @@ require (
 	gopkg.in/yaml.v2 v2.4.0
 )
 
-require go.bug.st/testsuite v0.0.1
+require (
+	go.bug.st/testifyjson v1.0.0
+	go.bug.st/testsuite v0.1.0
+)
 
 require (
 	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
