+++ internal/integrationtest/core/core_list_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/arduino/arduino-cli/internal/integrationtest"
 	"github.com/arduino/go-paths-helper"
 	"github.com/stretchr/testify/require"
-	"go.bug.st/testsuite/requirejson"
+	"go.bug.st/testifyjson/requirejson"
 )
 
 func TestCorrectHandlingOfPlatformVersionProperty(t *testing.T) {
